+++ Tests/Get-RubrikProxySetting.Tests.ps1
@@ -1,0 +1,43 @@
+Remove-Module -Name 'Rubrik' -ErrorAction 'SilentlyContinue'
+Import-Module -Name './Rubrik/Rubrik.psd1' -Force
+
+foreach ( $privateFunctionFilePath in ( Get-ChildItem -Path './Rubrik/Private' | Where-Object extension -eq '.ps1').FullName  ) {
+    . $privateFunctionFilePath
+}
+
+Describe -Name 'Public/Get-RubrikProxySetting' -Tag 'Public', 'Get-RubrikProxySetting' -Fixture {
+    #region init
+    $global:rubrikConnection = @{
+        id      = 'test-id'
+        userId  = 'test-userId'
+        token   = 'test-token'
+        server  = 'test-server'
+        header  = @{ 'Authorization' = 'Bearer test-authorization' }
+        time    = (Get-Date)
+        api     = 'v1'
+        version = '4.0.5'
+    }
+    #endregion
+
+    Context -Name 'Returned Results' {
+        Mock -CommandName Test-RubrikConnection -Verifiable -ModuleName 'Rubrik' -MockWith { }
+        Mock -CommandName Submit-Request -Verifiable -ModuleName 'Rubrik' -MockWith {
+            @{ 
+                'host'     = 'proxy.server.com' 
+                'protocol' = 'https'
+            }
+        }
+        It -Name 'No parameters returns all results' -Test {
+            @( Get-RubrikProxySetting | Measure-Object).Count |
+                Should -BeExactly 1
+        }
+        It -Name 'Host is proxy.server.com' -Test {
+            @( Get-RubrikProxySetting).host |
+                Should -BeExactly 'proxy.server.com'
+        }
+
+        Assert-VerifiableMock
+        Assert-MockCalled -CommandName Test-RubrikConnection -ModuleName 'Rubrik' -Exactly 2
+        Assert-MockCalled -CommandName Submit-Request -ModuleName 'Rubrik' -Exactly 2
+    }
+}